refactor: move renovate init logic to linter module Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
