+++ src/init/hooks.rs
@@ -1,4 +1,3 @@
-pub(crate) mod renovate_deps;
 pub(crate) mod rumdl;
 pub(crate) mod rustfmt;
 pub(crate) mod taplo;
